Add ItemController and refactor item MVC
